手贱删库了= =

+++ src/main/java/com/zerodeng/bean/system/SystemModules.java
@@ -27,9 +27,9 @@ public class SystemModules {
 
     private Date modifyTime;
 
-    private Integer createUser;
+    private Long createUser;
 
-    private Integer modifyUser;
+    private Long modifyUser;
 
     public Long getId() {
         return id;
@@ -127,19 +127,19 @@ public class SystemModules {
         this.modifyTime = modifyTime;
     }
 
-    public Integer getCreateUser() {
+    public Long getCreateUser() {
         return createUser;
     }
 
-    public void setCreateUser(Integer createUser) {
+    public void setCreateUser(Long createUser) {
         this.createUser = createUser;
     }
 
-    public Integer getModifyUser() {
+    public Long getModifyUser() {
         return modifyUser;
     }
 
-    public void setModifyUser(Integer modifyUser) {
+    public void setModifyUser(Long modifyUser) {
         this.modifyUser = modifyUser;
     }
 }
